#6 Used v0 dev to generate the API using this prompt:
I want to generate an application for importing vocabulary language with a text field that lets us import a thematic category for generating language vocabulary.When submitting a thematic category, the app should hit an API endpoint (API route in app router) to invoke LLM chat completions in Grog (LLM) on the server side and then pass that information back to the front end. The tech stack is:- Gradio- Streamlit- FastHTMLIt has to create a structured JSon is:[{"kanji":"","romaji":"","english":"","parts":[{"kanji":"","romaji":""},{"kanji":"","romaji":""}]}]We want to present this json in a textfield, it should be copy friendly and have a download button to download only the content from the json textfield.

Then proceed to set up my groq api in a .env_local and also modify the prompt given to groq to add an example (that I got from Chat GPT) as well as the groq model which was failing